requirements.txt
Flask==0.12.1
Flask-Migrate==2.0.3
Flask-SQLAlchemy==2.2
Flask-Script==2.0.5
Flask-WTF==0.14.2
Jinja2==2.9.6
Mako==1.0.6
MySQL-python==1.2.5
SQLAlchemy==1.1.9
WTForms==2.1
Werkzeug==0.12.1
alembic==0.9.1
argparse==1.4.0
itsdangerous==0.24
wsgiref==0.1.2
python-dotenv